requirements.txt
streamlit
google.generativeai
langchain_community 
PyPDF2
together